backends/pixi-build-ros/pyproject.toml
[project]
dependencies = ["rosdistro", "catkin_pkg", "pytest", "toml", "pyyaml"]
name = "pixi-build-ros"
version = "0.1.0"

[project.scripts]
pixi-build-ros = "pixi_build_ros.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pixi_build_ros"]

[tool.hatch.build.targets.wheel.shared-data]
"templates" = "site-packages/pixi_build_ros/templates"
"robostack.yaml" = "site-packages/pixi_build_ros/robostack.yaml"

[tool.hatch.build.targets.editable]
packages = ["src/pixi_build_ros"]

[tool.hatch.build.targets.editable.shared-data]
"templates" = "site-packages/pixi_build_ros/templates"
"robostack.yaml" = "site-packages/pixi_build_ros/robostack.yaml"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
